phase-2/to-do/backend/tests/api/test_auth.py
# backend/tests/api/test_auth.py
from fastapi.testclient import TestClient
from src.main import app # Assuming app is imported from main
from src.db import get_session # Assuming get_session is defined in db.py
import pytest

client = TestClient(app)

# Mock database session for testing
@pytest.fixture(name="session")
def session_fixture():
    # This is a placeholder for a proper test database setup
    # In a real scenario, you'd use an in-memory DB or a separate test DB
    yield get_session() # This will need to be properly mocked later

def test_signup_user():
    response = client.post(
        "/api/auth/signup",
        json={"email": "test@example.com", "password": "TestPassword123"}
    )
    assert response.status_code == 201
    assert "user_id" in response.json()

def test_signup_existing_user():
    # First signup
    client.post(
        "/api/auth/signup",
        json={"email": "existing@example.com", "password": "TestPassword123"}
    )
    # Try to signup again with same email
    response = client.post(
        "/api/auth/signup",
        json={"email": "existing@example.com", "password": "TestPassword123"}
    )
    assert response.status_code == 409 # Conflict

def test_login_user():
    # Assuming a user is already signed up for login test
    # This would be part of a proper test setup
    client.post(
        "/api/auth/signup",
        json={"email": "login@example.com", "password": "LoginPassword123"}
    )
    response = client.post(
        "/api/auth/login",
        json={"email": "login@example.com", "password": "LoginPassword123"}
    )
    assert response.status_code == 200
    assert "access_token" in response.json()
    assert response.json()["token_type"] == "bearer"

def test_login_incorrect_password():
    response = client.post(
        "/api/auth/login",
        json={"email": "login@example.com", "password": "WrongPassword"}
    )
    assert response.status_code == 401 # Unauthorized

def test_logout_user():
    # Placeholder for logout test, will require a valid token
    response = client.post(
        "/api/auth/logout",
        headers={"Authorization": "Bearer some_valid_token"} # This token would come from a successful login
    )
    assert response.status_code == 200
    assert response.json() == {"message": "Logged out successfully"}
