phase-2/to-do/backend/requirements.txt
fastapi
uvicorn
sqlmodel
psycopg2-binary
passlib
bcrypt
python-jose
python-dotenv
